Move into the sneaky folder.
