+++ apps/dotcom/client/src/tla/components/TlaEditor/TlaPublishEditor.tsx
@@ -6,11 +6,11 @@ import { useLegacyUrlParams } from '../../../hooks/useLegacyUrlParams'
 import { useHandleUiEvents } from '../../../utils/analytics'
 import { assetUrls } from '../../../utils/assetUrls'
 import { globalEditor } from '../../../utils/globalEditor'
-import { SneakyDarkModeSync } from './SneakyDarkModeSync'
 import { TlaEditorTopLeftPanel } from './TlaEditorTopLeftPanel'
 import { TlaEditorErrorFallback } from './editor-components/TlaEditorErrorFallback'
 import { TlaEditorPublishedSharePanel } from './editor-components/TlaEditorPublishedSharePanel'
 import styles from './editor.module.css'
+import { SneakyDarkModeSync } from './sneaky/SneakyDarkModeSync'
 import { useFileEditorOverrides } from './useFileEditorOverrides'
 
 const components: TLComponents = {
