+++ apps/dotcom/client/src/tla/pages/local.tsx
@@ -4,8 +4,8 @@ import { assert, react } from 'tldraw'
 import { LocalEditor } from '../../components/LocalEditor'
 import { routes } from '../../routeDefs'
 import { globalEditor } from '../../utils/globalEditor'
-import { SneakyDarkModeSync } from '../components/TlaEditor/SneakyDarkModeSync'
 import { components } from '../components/TlaEditor/TlaEditor'
+import { SneakyDarkModeSync } from '../components/TlaEditor/sneaky/SneakyDarkModeSync'
 import { useMaybeApp } from '../hooks/useAppState'
 import { TlaAnonLayout } from '../layouts/TlaAnonLayout/TlaAnonLayout'
 import { clearShouldSlurpFile, getShouldSlurpFile, setShouldSlurpFile } from '../utils/slurping'
